extracting the test fakes to their own package
